Sorted files, Changed few lines of code
Sorted files into some folders, that should be easy to understand for a normal human. Changed few lines of code. Removed useless line and changed it with slightly simpler alternative. Changed the name of one variable to also something more readable and reasonable. Added the comment (mostly for myself) regarding the name of this script. Please use "...Manager" next time...