Made basic "room" spawning
Made new file called "Rooms" where we will place all the rooms in the game. Obviously we will later sort it out on layers so it wont be randomly generated chaos.

Besides that, added basic script "DungeonGenerator" and a empty game object called "Game manager". The script has a safety sort of, where if it cant spawn a room after 50 tries, it increases the size of spawning arena, that ensures decent room layouts. Obviously if player will get EXTREMELY LUCKY, then one room will spawn extremely far away. But chances for that are less than zero or so.

Added few basic red rooms that ressebmle normal rooms. They lack doors and their main function for now is a placeholder.

The way that the script works is:
1. Spawn a room from a array
2. Remove that element from array
3. is this room colliding with others?
4. if yes, try to spawn it again
5. else, break

That should be it for now. i will expand upon this later but it shouldnt be much of a struggle. (BESIDES THOSE NERDY ALGORITHMS).