"Fixed" the rooms
Just changed the doors to be more "grid" like, changed the scale of the doors to be pretty rounded up (x: 0.6, y: 0.6, z: 0.1). It will make it easier to add corridors later on. We still need to somehow know where the door is facing, and later on to tell the amount of doors.