More readable code, attempt at fixing
Made code more readable, basically added few more spaces, and moved WHOLE start function to the new one called MakeRooms().

Tried to add some type of logic into room connections, but failed. Wanted to do something (if the rooms is connected, ignore it) but failed, because the "not set instance to an object" error kept popping out on the line 171 (which is the longest one, and i couldnt understand that much) . Hopefully YOU understand the code a little bit, and i hope it will be fixed, cuz i dont understand the line 171, and overall the class edge. Cya!